u128 not supported for config

+++ committer/src/config.rs
@@ -69,7 +69,7 @@ pub struct Eth {
     pub state_contract_address: Address,
     /// This is a workaround to get pending transactions unstuck until the tx manager task is
     /// complete.
-    pub max_fee_per_gas_for_first_tx: Option<u128>,
+    pub max_fee_per_gas_for_first_tx: Option<u64>,
 }
 
 fn parse_url<'de, D>(deserializer: D) -> Result<Url, D::Error>
